backend/requirements.txt
Django
djangorestframework
pylint
